requirements.txt
debugpy==1.8.1
fastapi==0.111.0
gunicorn==22.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-logging==0.45b0
pre-commit==3.7.1
prometheus-client==0.20.0
torch==2.3.0
transformers==4.41.1
